+++ Package.swift
@@ -1,4 +1,4 @@
-// swift-tools-version:5.10
+// swift-tools-version:6.0
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
@@ -16,11 +16,18 @@ let package = Package(
     ],
     targets: [
         .target(
-            name: "Biodag"
+            name: "Biodag",
+            swiftSettings: [
+                .swiftLanguageMode(.v6)
+            ]
         ),
         .testTarget(
             name: "BiodagTests",
-            dependencies: ["Biodag"]
+            dependencies: ["Biodag"],
+            swiftSettings: [
+                .swiftLanguageMode(.v6)
+            ]
         )
-    ]
+    ],
+    swiftLanguageModes: [.v6]
 )